aws_lambda_capacity_provider: separate file
